Split main into lib and run modules
