+++ src/lib.rs
@@ -1,0 +1,11 @@
+extern crate alloc;
+
+#[cfg(feature = "profiler")]
+extern crate cpuprofiler;
+
+pub mod run;
+
+mod configuration;
+mod platform;
+mod util;
+mod wireguard;
